refactor: move cli utils to deployment module
