+++ cli/src/main/java/io/quarkiverse/backstage/cli/template/InstallCommand.java
@@ -6,8 +6,8 @@ import java.util.List;
 import java.util.Optional;
 
 import io.quarkiverse.backstage.cli.common.GenerationBaseCommand;
-import io.quarkiverse.backstage.cli.utils.Git;
-import io.quarkiverse.backstage.cli.utils.Projects;
+import io.quarkiverse.backstage.deployment.utils.Git;
+import io.quarkiverse.backstage.deployment.utils.Projects;
 import io.quarkiverse.backstage.rest.CreateLocationRequest;
 import io.quarkiverse.backstage.rest.RefreshEntity;
 import io.quarkiverse.backstage.runtime.BackstageClient;
