+++ deployment/src/main/java/io/quarkiverse/backstage/deployment/utils/Directories.java
@@ -1,4 +1,4 @@
-package io.quarkiverse.backstage.cli.utils;
+package io.quarkiverse.backstage.deployment.utils;
 
 import java.io.IOException;
 import java.nio.file.Files;
