+++ pom.xml
@@ -38,8 +38,7 @@
     <sundrio.version>0.101.0</sundrio.version>
     <lombok.version>1.18.26</lombok.version>
     <rs-api.version>3.1.0</rs-api.version>
-    <quarkus-jgit.version>3.1.2</quarkus-jgit.version>
-    <quarkus-github-api.version>1.322.0</quarkus-github-api.version>
+    <quarkus-jgit.version>3.2.2</quarkus-jgit.version>
     <surefire-plugin.version>3.2.5</surefire-plugin.version>
   </properties>
